backend/requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
djangorestframework==3.9.1
django-rest-auth==0.9.3
django-allauth==0.38.0
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.2.1
django-heroku==0.3.1
gunicorn==19.9.0